requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
